Revert "refactor: use hashmap instead of qdata on saving handlers in SongList and Recordings"
This reverts commit 1111e0f26dd81f42e7b495f61569931e9ffed5eb.

It was problematic since not all id is mapped to a single instance.

In this case, since we bind first to the new song before unbinding the previous
one, the handler id of the previous song is overwritten with the handler id of the
new song. This results to unbinding the previous song from the handler id of the
new song, which causes a critical error.

A way to fix this is to unbind the previous song, before binding the new one, but
that is prone to error.